move routes out of netlink.go
